+++ tests/shipping/models/Delivery_Scheduled Delivery - Credit card.model.js
@@ -13,7 +13,7 @@ import {
   completePurchase,
   payWithCreditCard,
 } from "../../../utils/payment-actions"
-import { ACCOUNT_NAMES } from "../../../utils/constants"
+import { SKUS } from "../../../utils/constants"
 
 export default function test(account) {
   describe(`Delivery + Scheduled Delivery - Credit card - ${account}`, () => {
@@ -24,7 +24,7 @@ export default function test(account) {
     it("delivery with scheduled delivery with multiple items", () => {
       const email = getRandomEmail()
 
-      setup({ skus: ["35", "291"], account })
+      setup({ skus: [SKUS.DELIVERY_AND_PICKUP, SKUS.SCHEDULED_DELIVERY], account })
 
       fillEmail(email)
       fillProfile()