+++ tests/shipping/models/Delivery_Scheduled Delivery - Second Purchase - Credit card.model.js
@@ -9,7 +9,7 @@ import {
   chooseDeliveryDate,
 } from "../../../utils/shipping-actions"
 import { completePurchase, typeCVV } from "../../../utils/payment-actions"
-import { ACCOUNT_NAMES } from "../../../utils/constants"
+import { SKUS } from "../../../utils/constants"
 
 export default function test(account) {
   describe(`Delivery + Scheduled Delivery - 2P - Credit card - ${account}`, () => {
@@ -20,7 +20,10 @@ export default function test(account) {
     it("delivery with scheduled delivery with multiple items", () => {
       const email = getSecondPurchaseEmail()
 
-      setup({ skus: ["35", "291"], account })
+      setup({
+        skus: [SKUS.DELIVERY_AND_PICKUP, SKUS.SCHEDULED_DELIVERY],
+        account,
+      })
       fillEmail(email)
       confirmSecondPurchase()
       chooseDeliveryDate({ account })