+++ tests/shipping/models/Delivery_Scheduled Delivery and Delivery - Credit card.model.js
@@ -13,7 +13,7 @@ import {
   completePurchase,
   payWithCreditCard,
 } from "../../../utils/payment-actions"
-import { ACCOUNT_NAMES } from "../../../utils/constants"
+import { SKUS } from "../../../utils/constants"
 
 export default function test(account) {
   describe(`Delivery + Scheduled Delivery and Delivery - Credit card - ${account}`, () => {
@@ -24,7 +24,13 @@ export default function test(account) {
     it("one item with delivery and another item with both scheduled delivery and delivery", () => {
       const email = getRandomEmail()
 
-      setup({ skus: ["35", "299"], account })
+      setup({
+        skus: [
+          SKUS.DELIVERY_AND_PICKUP,
+          SKUS.SCHEDULED_DELIVERY_AND_DELIVERY_MULTIPLE_SLA,
+        ],
+        account,
+      })
 
       fillEmail(email)
       fillProfile()