+++ tests/shipping/models/Pickup_Delivery - Second Purchase - Credit card.model.js
@@ -12,7 +12,7 @@ import {
 } from "../../../utils/shipping-actions"
 import { completePurchase, typeCVV } from "../../../utils/payment-actions"
 import { goToInvoiceAddress } from "../../../utils/invoice-actions"
-import { ACCOUNT_NAMES } from "../../../utils/constants"
+import { ACCOUNT_NAMES, SKUS } from "../../../utils/constants"
 
 export default function test(account) {
   describe(`Pickup + Delivery - 2P - Credit card - ${account}`, () => {
@@ -23,7 +23,7 @@ export default function test(account) {
     it("with only pickup", () => {
       const email = getSecondPurchaseEmail()
 
-      setup({ skus: ["285", "289"], account })
+      setup({ skus: [SKUS.PICKUP_1_SLA, SKUS.DELIVERY_MULTIPLE_SLA], account })
       fillEmail(email)
       confirmSecondPurchase()
       unavailableDeliveryGoToPickup()