+++ tests/shipping/models/Pickup_Scheduled Delivery - Second Purchase - Credit card.model.js
@@ -13,7 +13,7 @@ import {
 } from "../../../utils/shipping-actions"
 import { completePurchase, typeCVV } from "../../../utils/payment-actions"
 import { goToInvoiceAddress } from "../../../utils/invoice-actions"
-import { ACCOUNT_NAMES } from "../../../utils/constants"
+import { ACCOUNT_NAMES, SKUS } from "../../../utils/constants"
 
 export default function test(account) {
   describe(`Pickup + Scheduled Delivery - 2P - Credit card - ${account}`, () => {
@@ -24,7 +24,10 @@ export default function test(account) {
     it("with only pickup", () => {
       const email = getSecondPurchaseEmail()
 
-      setup({ skus: ["285", "291"], account })
+      setup({
+        skus: [SKUS.PICKUP_1_SLA, SKUS.SCHEDULED_DELIVERY],
+        account,
+      })
       fillEmail(email)
       confirmSecondPurchase()
       unavailableDeliveryGoToPickup()